Rename and some fixes
